Node system: get rid of reinterpret_cast
NOTE: previous version depended on assumption that
reinteret_cast<shared_ptr<U>>(shared_ptr<T>) is safe, which, while
probably being true for all relevant implementations in relevant
cases, is not guranteed by standard. This commit replaces
reinterpret_cast with a couple of smart pointer casts, which might
be considerably slower. If that will be the case, we might need to
revert to unsafe version or think up something else.